FindBugs: Whole bunch of smaller modifications to make FindBugs happy.
E.g.:
Integer division assigned to double;
Appending to string in loop; unnecessary use of new String();
Possibly stream not closed;
Variable and method names should start lowercase;
Should not do double comparison with "==";
Make final arrays private and provide getters, instead of havbe them public;
Inefficient use of Map iterator (use EntrySet insteaf of KeySet and lookup);
read (from random device, version file) might not read as much as expected 
  - check for that;
Some static never modified variable now final static.

Most of those changes should not have any effect at all, but the double == 0
and integer divisions (which were probably meant to be double divisions)
might effect e.g. AI calculations.


git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@3479 c5679a2a-0e0e-0410-bba2-8642b794b8c1

+++ Colossus/core/src/main/java/net/sf/colossus/ai/RationalAI.java
@@ -2207,8 +2207,8 @@ public class RationalAI extends SimpleAI
 
         double expectedValue;
 
-        expectedValue = defenderKilled / numOtherPlayers - attackerKilled
-            + attackerMuster - defenderMuster / numOtherPlayers;
+        expectedValue = defenderKilled / (double)numOtherPlayers - attackerKilled
+            + attackerMuster - defenderMuster / (double)numOtherPlayers;
 
         if (attackerCreatures.size() > 1)
         {
